refactor: cleanup and refactor into "proper" contexts for Images and Tags
