+++ apps/onagal/lib/onagal/tags/gallery.ex
@@ -1,4 +1,4 @@
-defmodule Onagal.Gallery do
+defmodule Onagal.Tags.Gallery do
   use Ecto.Schema
   import Ecto.Changeset
 
@@ -9,7 +9,7 @@ defmodule Onagal.Gallery do
 
     ## TODO: on_replace may be wrong
     # , on_replace: :delete
-    many_to_many(:tags, Onagal.Tag, join_through: "galleries_tags")
+    many_to_many(:tags, Onagal.Tags.Tag, join_through: "galleries_tags")
   end
 
   @valid_params ~w(name)a
