+++ apps/onagal/lib/onagal/tags/tag.ex
@@ -1,4 +1,4 @@
-defmodule Onagal.Tag do
+defmodule Onagal.Tags.Tag do
   use Ecto.Schema
   import Ecto.Changeset
 
@@ -9,9 +9,9 @@ defmodule Onagal.Tag do
 
     ## TODO: on_replace may be wrong
     # , on_replace: :delete
-    many_to_many(:images, Onagal.Image, join_through: "images_tags")
+    many_to_many(:images, Onagal.Images.Image, join_through: "images_tags")
     # , on_replace: :delete
-    many_to_many(:tagsets, Onagal.Tagset, join_through: "tags_tagsets")
+    many_to_many(:tagsets, Onagal.Tags.Tagset, join_through: "tags_tagsets")
   end
 
   @doc false
